Applied hotfix patch to allow proper api responses for android app development
